Added initial watcher unit tests Added initial template tests.
