+++ lib/watcher.js
@@ -5,6 +5,11 @@ const chokidar = require('chokidar');
  * Class to watch for change in certain file(s)
  */
 class Watcher {
+
+  /**
+   * Watch over given paths.
+   * @param {*} paths to watch
+   */
   constructor(paths) {
     if (Array.isArray(paths)) {
       this.paths = paths;
