+++ test/integration/templates/java-spring.spec.js
@@ -4,23 +4,8 @@ const options = { compareSize: true };
 const pathToExpectedTemplates = './test/templates/java-spring';
 const pathToactualTemplates = './test/integration/templates/actual/java-spring';
 const { exec } = require('child_process');
-const fs = require('fs');
-const Path = require('path');
-const format = require('util').format;
-var expect = require('chai').expect;
-const deleteFolderRecursive = function(path) {
-  if (fs.existsSync(path)) {
-    fs.readdirSync(path).forEach((file, index) => {
-      const curPath = Path.join(path, file);
-      if (fs.lstatSync(curPath).isDirectory()) { // recurse
-        deleteFolderRecursive(curPath);
-      } else { // delete file
-        fs.unlinkSync(curPath);
-      }
-    });
-    fs.rmdirSync(path);
-  }
-};
+const expect = require('chai').expect;
+const {deleteFolderRecursive} = require('../utils/utils');
 describe('java-spring', () => {
   before(() => {
     deleteFolderRecursive(pathToactualTemplates);
@@ -34,12 +19,12 @@ describe('java-spring', () => {
       });
       child.on('exit', (code) => {
         if (code === 0) {
-          const res = dircompare.compareSync(pathToactualTemplates+"/no-params", `${pathToExpectedTemplates}/no-params`, options);
-          try{
+          const res = dircompare.compareSync(`${pathToactualTemplates}/no-params`, `${pathToExpectedTemplates}/no-params`, options);
+          try {
             expect(res.same).to.be.true;
             resolve();
             deleteFolderRecursive(pathToactualTemplates);
-          }catch(e){
+          }catch (e) {
             reject(e);
           }
         } else {
